requirements.txt
boto3
dnspython
sslyze>=5.0,<6.0
